+++ Runtime/Levels/LevelData.cs
@@ -1,12 +1,11 @@
 using UnityEngine;
 using UnityEngine.AddressableAssets;
-using UnityEngine.Rendering;
 
 namespace Pixygon.Micro
 {
     [CreateAssetMenu(menuName = "Pixygon/Micro/New LevelData")]
     public class LevelData : ScriptableObject {
-        //public GameObject _levelPrefab;
+        public string _levelName;
         public AssetReference _levelRef;
         public AssetReference _bgmRef;
         public AssetReference _postProcessingProfileRef;