+++ Runtime/MicroActor/MicroActorData.cs
@@ -8,6 +8,7 @@ namespace Pixygon.Micro {
         public AssetReference _actorRef;
         public float _speed;
         public int _hp;
+        public int _pointsValue;
         public bool _isHostile;
         public bool _isKillable;
         public bool _detectGaps;